+++ src/Blade.php
@@ -5,19 +5,19 @@ namespace Surgiie\Blade;
 use Closure;
 use Illuminate\Container\Container;
 use Illuminate\Contracts\Foundation\Application as FoundationApplication;
+use Illuminate\Contracts\View\Factory as ViewFactoryContract;
 use Illuminate\Events\Dispatcher;
 use Illuminate\Filesystem\Filesystem;
 use Illuminate\View\Engines\EngineResolver;
+use Illuminate\View\ViewFinderInterface;
 use SplFileInfo;
+use Surgiie\Blade\Concerns\ParsesFilePath;
 use Surgiie\Blade\Exceptions\FileNotFoundException;
 use Surgiie\Blade\Exceptions\UndefinedVariableException;
 
 class Blade
 {
-    /**
-     * The render file instance.
-     */
-    protected ?File $file = null;
+    use ParsesFilePath;
 
     /**
      * Get the engine name for resolver registration.
@@ -35,14 +35,17 @@ class Blade
     protected ?SplFileInfo $fileInfo = null;
 
     /**
+     * The file finder instance.
+     */
+    protected ?FileFinder $fileFinder = null;
+
+    /**
      * The file factory instance.
      */
     protected ?FileFactory $fileFactory = null;
 
-    /**
-     * The file finder instance.
-     */
-    protected ?FileFinder $fileFinder = null;
+    /**Whether cached files should get used.*/
+    protected static bool $useCachedFiles = true;
 
     /**
      * The engine resolver instance.
@@ -70,8 +73,21 @@ class Blade
     public function __construct(Container|FoundationApplication $container, Filesystem $filesystem)
     {
         $this->container = $container;
-
         $this->filesystem = $filesystem;
+
+        $this->container->bind(ViewFactoryContract::class, function () {
+            return $this->getFileFactory();
+        });
+
+        $this->container->bind('view', function () {
+            return $this->getFileFactory();
+        });
+
+        $this->container->bind(ViewFinderInterface::class, function () {
+            return $this->getFileFinder();
+        });
+
+        Container::setInstance($this->container);
 
         $this->makeCompiledDirectory();
 
@@ -80,6 +96,18 @@ class Blade
         $this->resolver->register(self::ENGINE_NAME, function () {
             return $this->getCompilerEngine();
         });
+    }
+
+    /**Set whether cached files should be used or not. */
+    public static function useCachedCompiledFiles(bool $useCacheFiles)
+    {
+        static::$useCachedFiles = $useCacheFiles;
+    }
+
+    /**Get whether cached compiled files should be used or not. */
+    public static function shouldUseCachedCompiledFiles()
+    {
+        return static::$useCachedFiles;
     }
 
     /**Normalize a path for the appropriate OS/directory separator.*/
@@ -159,7 +187,7 @@ class Blade
     /**
      * Get the compiled path to where compiled files go.
      */
-    protected function getCompiledPath(): string
+    public function getCompiledPath(): string
     {
         return __DIR__.'/../.compiled';
     }
@@ -196,15 +224,24 @@ class Blade
     {
         $path = static::normalizePathForOS($path);
 
-        if (! is_file($path)) {
+        $path = static::parseFilePath($path);
+
+        $real_path = realpath($path);
+
+        if ($real_path === false || ! is_file($path)) {
             throw new FileNotFoundException("The $path file does not exist.");
         }
+        $finder = $this->getFileFinder();
 
-        $info = new SplFileInfo($path);
+        $finder->replaceNamespace('__components', $this->getCompiledPath());
+
+        $info = new SplFileInfo($real_path);
 
         $factory = $this->getFileFactory();
+        // flush found files, so we're not returning files that match in path when using relative paths.
+        $finder->flush();
 
-        $this->getFileFinder()->setPaths([dirname($info->getRealPath())]);
+        $finder->setPaths([dirname($info->getRealPath())]);
 
         $factory->addExtension($info->getExtension(), self::ENGINE_NAME);