+++ src/FileCompilerEngine.php
@@ -10,8 +10,9 @@ class FileCompilerEngine extends CompilerEngine
 {
     /**
      * Overwritten to not ltrim but to rtrim outbutput buffer.
-     *
-     * This assists with preserving spacing/indentation from compiled file.
+     * This assists with preserving spacing/indentation from
+     * the compiled file which is important to preserve so
+     * we dont end up with shifted content in a file.
      */
     protected function evaluatePath($path, $data)
     {
@@ -37,7 +38,6 @@ class FileCompilerEngine extends CompilerEngine
     protected function handleViewException(Throwable $e, $obLevel)
     {
         $class = get_class($e);
-
         PhpEngine::handleViewException(new $class($this->getMessage($e)), $obLevel);
     }